Move on to JSON output for pitch log
I decided to move on from the TUI version of the game for now. I was
spending way too much time creating a TUI when the end goal is a web
app. For now, I'll serialize the pitch log data type into JSON and use
common tools for inspecting the data. Then I should focus on fixing the
game logic.